remove duplicate task
